requirements.txt
accelerate
diffusers
fire
ipython
jupyter
numpy
peft
Pillow
torchvision
torch
transformers
